Follow fix (#71)
* Updated the seeds of the artists and genres and changed 2 endpoints to cope up with the new change

* Modified the seeds of tracks and albums

* Add more artists, genres

* Add adele emeinem seeds

* Add seeds

* Added the last artist in the seeds

* Update artists.json

* Update artists.json

* Update artists.json

* Fix Follow endpoints

* Fix Follow unit tests

+++ src/services/follow.service.js
@@ -203,6 +203,7 @@ exports.followUser = async (ids, type, user) => {
 exports.unfollowUser = async (ids, type, user) => {
   ids = _.uniq(ids);
   const users = await User.find({ _id: ids, type: type });
+  if(!users || users.length < ids.length) return false
   await Followings.deleteMany({
     userId: user._id,
     followedId: users,
